refactor: move GAME function handling to a separate file
